i2c + zforce: handle the small gap without interrupts during suspend gracefully
During suspend there is a small time, when interrupts are disabled and the
wakeup-sources are not yet active. When the zforce produces an interrupt
at exactly this moment it is not handled and the already low level of the
pin at this point also prevents the system from waking up, as the level
does not change for subsequent messages the touch controller wants to send.

Fix this by introducing suspend_noirq functions both to the i2c subsystem
and the zforce driver that check the level of the interrupt gpio at the
latest possible time and prevent the system from going to sleep if necessary.