lk_tps65185: fix race condition between workqueue and suspend
[first part about locking resolved upstream]

Secondly add pm_wakeup calls in _pwrdwn_work_func, to make the sure the
system does not suspend while pwrdwn_work_func is running, as this could
also lead to a deadlock. Tasks get frozen first when suspending, so when
_pwrdwn_work_func is running at the time, it is possible it will be stopped
and keep holding the lock, making the later _chg_mode calls during suspend
hang the system, waiting for the lock.