tps65185: allow deep suspend
Until now the tps65185 would always abort our deep suspend due to some
timeout for some voltage to settle. This is tied to some uiconfig option
from ntx and changing that timeout to 0 works just fine for our deep
suspend.
Additionally we convert the abortion of the suspend into a wait if we
at some point might happen to run into this timeout again.