Store the still unprocessed Q22 Patches as reference.
As the kernels are so very similar, it makes sense to check if we solved
particular problems in the previous ereaders already. I already ported
a lot of interesting patches, hence the missing numbers in the list.

Patches should not be ported without thinking of course, but can be a
help when some particular problem comes up. So store them here as reference.

When applying a patch from that list, please also remove it from here,
so to keep track of what is still unused.