don't manually enable/disable rtc interrupt in suspend/resume
The rtc interrupt handling was a bit over-engineered in that it tried
tracking if the alarm was user-enabled and then enabled/disabled the
interrupt manually in the suspend/resume callbacks of the ricoh619.

But the interrupt already gets enabled / disabled correctly when the
alarm itself is enabled / disabled, so there is no need for that.
Additionally the flag got reset on resume, thus in our multiple
suspend tries, only the first one would get the interrupt and thus
the rtc-wake-flag tracking. A second resume from suspend would then
look like a manual resume without rtc involvement.